refactor(hl7): export getFreePort from @medplum/hl7 as single impl
The flaky-port fix previously lived duplicated in both the agent and hl7
test-utils. Consolidate it into one implementation in
packages/hl7/src/free-port.ts, exported from the package's public API, and
have both packages import it directly:

- hl7's own tests import from './free-port'
- the agent package imports from '@medplum/hl7'

free-port.ts is a standalone, vitest-free module so it can be exported from
the runtime index without pulling test-only globals into the bundle.
