+++ packages/agent/src/test-utils.ts
@@ -3,9 +3,9 @@
 import type { ILogger, MedplumClient, WithId } from '@medplum/core';
 import { LogLevel, TypedEventTarget, sleep } from '@medplum/core';
 import type { Endpoint } from '@medplum/fhirtypes';
+import { getFreePort } from '@medplum/hl7';
 import { randomUUID } from 'node:crypto';
 import { mkdirSync, rmSync } from 'node:fs';
-import { createServer } from 'node:net';
 import os from 'node:os';
 import path from 'node:path';
 import type { Mock } from 'vitest';
@@ -108,49 +108,6 @@ export function createTestHl7ClientPool(
   const heartbeatEmitter = options.heartbeatEmitter ?? new TypedEventTarget();
   const pool = new Hl7ClientPool({ ...options, messageTracker, heartbeatEmitter });
   return { pool, messageTracker, heartbeatEmitter };
-}
-
-// Every port ever handed out by `getFreePort`, so we never issue the same one twice
-// within a single test process — see the comment on `getFreePort` below.
-const issuedPorts = new Set<number>();
-
-// Used only for tests that need a free port number with *nothing* listening on it.
-// For tests that start an Hl7Server, prefer `server.start(0)` which returns the OS-assigned
-// port and never has a release-then-rebind window.
-//
-// Because we close the probing server before returning, the port is freed immediately and the
-// OS may hand the *same* ephemeral port to a subsequent call. Callers that allocate two ports
-// (e.g. one per channel) would then collide. To avoid this, remember every port we issue and
-// keep any probing server that lands on an already-issued port open until we find a fresh one,
-// so the OS can't reissue it in the same round.
-export async function getFreePort(): Promise<number> {
-  const heldServers: ReturnType<typeof createServer>[] = [];
-  const closeServer = async (server: ReturnType<typeof createServer>): Promise<void> =>
-    new Promise((resolve, reject) => {
-      server.close((err) => (err ? reject(err) : resolve()));
-    });
-  try {
-    for (let attempt = 0; attempt < 20; attempt++) {
-      const server = createServer();
-      const port = await new Promise<number>((resolve, reject) => {
-        server.on('error', reject);
-        server.listen(0, () => {
-          resolve((server.address() as { port: number }).port);
-        });
-      });
-      if (issuedPorts.has(port)) {
-        // Keep this server listening so the OS won't hand the same port out again this round.
-        heldServers.push(server);
-        continue;
-      }
-      issuedPorts.add(port);
-      await closeServer(server);
-      return port;
-    }
-    throw new Error('Unable to find a free port after 20 attempts');
-  } finally {
-    await Promise.all(heldServers.map((server) => closeServer(server).catch(() => undefined)));
-  }
 }
 
 export async function createEndpointWithRandomPort(
